+++ lib/djvGui/djvFloatEdit.h
@@ -56,7 +56,7 @@ class DJV_GUI_EXPORT djvFloatEdit : public QWidget
     
     //! This property holds the value.
     Q_PROPERTY(
-        double value
+        float  value
         READ   value
         WRITE  setValue
         NOTIFY valueChanged)
@@ -67,42 +67,42 @@ public:
     virtual ~djvFloatEdit();
 
     //! Get the value.
-    double value() const;
+    float value() const;
 
     //! Get the minimum value.
-    double min() const;
+    float min() const;
 
     //! Get the maximum value.
-    double max() const;
+    float max() const;
 
     //! Get the floating-point object.
     djvFloatObject * object() const;
 
 public Q_SLOTS:
     //! Set the value.
-    void setValue(double);
+    void setValue(float);
 
     //! Set the minimum value.
-    void setMin(double);
+    void setMin(float);
 
     //! Set the maximum value.
-    void setMax(double);
+    void setMax(float);
 
     //! Set the value range.
-    void setRange(double min, double max);
+    void setRange(float min, float max);
     
 Q_SIGNALS:
     //! This signal is emitted when the value is changed.
-    void valueChanged(double);
+    void valueChanged(float);
 
     //! This signal is emitted when the minimum value is changed.
-    void minChanged(double);
+    void minChanged(float);
 
     //! This signal is emitted when the maximum value is changed.
-    void maxChanged(double);
+    void maxChanged(float);
 
     //! This signal is emitted when the value range is changed.
-    void rangeChanged(double, double);
+    void rangeChanged(float, float);
     
 protected:
     virtual void keyPressEvent(QKeyEvent *);
