+++ lib/djvGui/djvFloatSlider.h
@@ -56,21 +56,21 @@ class DJV_GUI_EXPORT djvFloatSlider : public QWidget
     
     //! This property holds the value.    
     Q_PROPERTY(
-        double value
+        float  value
         READ   value
         WRITE  setValue
         NOTIFY valueChanged)
     
     //! This property holds the minimum value.
     Q_PROPERTY(
-        double min
+        float  min
         READ   min
         WRITE  setMin
         NOTIFY minChanged)
     
     //! This property holds the maximum value.
     Q_PROPERTY(
-        double max
+        float  max
         READ   max
         WRITE  setMax
         NOTIFY maxChanged)
@@ -81,42 +81,42 @@ public:
     virtual ~djvFloatSlider();
 
     //! Get the value.
-    double value() const;
+    float value() const;
 
     //! Get the minimum value.
-    double min() const;
+    float min() const;
 
     //! Get the maximum value.
-    double max() const;
+    float max() const;
 
     //! Get the floating-point object.
     djvFloatObject * object() const;
     
 public Q_SLOTS:
     //! Set the value.
-    void setValue(double);
+    void setValue(float);
 
     //! Set the minimum value.
-    void setMin(double);
+    void setMin(float);
 
     //! Set the maximum value.
-    void setMax(double);
+    void setMax(float);
 
     //! Set the value range.
-    void setRange(double min, double max);
+    void setRange(float min, float max);
 
 Q_SIGNALS:
     //! This signal is emitted when the value is changed.
-    void valueChanged(double);
+    void valueChanged(float);
 
     //! This signal is emitted when the minimum value is changed.
-    void minChanged(double);
+    void minChanged(float);
 
     //! This signal is emitted when the maximum value is changed.
-    void maxChanged(double);
+    void maxChanged(float);
 
     //! This signal is emitted when the range is changed.
-    void rangeChanged(double, double);
+    void rangeChanged(float, float);
     
 protected:
     virtual void keyPressEvent(QKeyEvent *);
